+++ buildstockbatch/base.py
@@ -19,12 +19,10 @@ import pandas as pd
 import requests
 import shutil
 import tempfile
-import yaml
 import yamale
 import zipfile
 import csv
 from collections import defaultdict, Counter
-import xml.etree.ElementTree as ET
 
 from buildstockbatch.__version__ import __schema_version__
 from buildstockbatch import (
@@ -33,7 +31,7 @@ from buildstockbatch import (
     postprocessing
 )
 from buildstockbatch.exc import SimulationExists, ValidationError
-from buildstockbatch.utils import path_rel_to_file
+from buildstockbatch.utils import path_rel_to_file, get_project_configuration
 
 logger = logging.getLogger(__name__)
 
@@ -56,7 +54,7 @@ class BuildStockBatchBase(object):
         self.project_filename = os.path.abspath(project_filename)
 
         # Load project file to self.cfg
-        self.cfg = self.get_project_configuration(project_filename)
+        self.cfg = get_project_configuration(project_filename)
 
         self.buildstock_dir = self.cfg['buildstock_directory']
         if not os.path.isdir(self.buildstock_dir):
@@ -208,31 +206,11 @@ class BuildStockBatchBase(object):
         assert(BuildStockBatchBase.validate_misc_constraints(project_file))
         assert(BuildStockBatchBase.validate_xor_nor_schema_keys(project_file))
         assert(BuildStockBatchBase.validate_reference_scenario(project_file))
-        assert(BuildStockBatchBase.validate_measures_and_arguments(project_file))
         assert(BuildStockBatchBase.validate_options_lookup(project_file))
         assert(BuildStockBatchBase.validate_measure_references(project_file))
         assert(BuildStockBatchBase.validate_options_lookup(project_file))
         logger.info('Base Validation Successful')
         return True
-
-    @classmethod
-    def get_project_configuration(cls, project_file):
-        try:
-            with open(project_file) as f:
-                cfg = yaml.load(f, Loader=yaml.SafeLoader)
-        except FileNotFoundError as err:
-            logger.error('Failed to load input yaml for validation')
-            raise err
-
-        # Set absolute paths
-        cfg['buildstock_directory'] = path_rel_to_file(project_file, cfg['buildstock_directory'])
-        if 'precomputed_sample' in cfg.get('baseline', {}):
-            cfg['baseline']['precomputed_sample'] = \
-                path_rel_to_file(project_file, cfg['baseline']['precomputed_sample'])
-        if 'weather_files_path' in cfg:
-            cfg['weather_files_path'] = path_rel_to_file(project_file, cfg['weather_files_path'])
-
-        return cfg
 
     @staticmethod
     def get_buildstock_dir(project_file, cfg):
@@ -244,7 +222,7 @@ class BuildStockBatchBase(object):
 
     @staticmethod
     def validate_sampler(project_file):
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
+        cfg = get_project_configuration(project_file)
         sampler_name = cfg['sampler']['type']
         try:
             Sampler = BuildStockBatchBase.get_sampler_class(sampler_name)
@@ -255,7 +233,7 @@ class BuildStockBatchBase(object):
 
     @staticmethod
     def validate_project_schema(project_file):
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
+        cfg = get_project_configuration(project_file)
         schema_version = cfg.get('schema_version')
         version_schema = os.path.join(os.path.dirname(__file__), 'schemas', f'v{schema_version}.yaml')
         if not os.path.isfile(version_schema):
@@ -268,7 +246,7 @@ class BuildStockBatchBase(object):
     @staticmethod
     def validate_misc_constraints(project_file):
         # validate other miscellaneous constraints
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
+        cfg = get_project_configuration(project_file)
 
         if cfg.get('postprocessing', {}).get('aggregate_timeseries', False):
             logger.warning('aggregate_timeseries has been deprecated and will be removed in a future version.')
@@ -277,7 +255,7 @@ class BuildStockBatchBase(object):
 
     @staticmethod
     def validate_xor_nor_schema_keys(project_file):
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
+        cfg = get_project_configuration(project_file)
         major, minor = cfg.get('version', __schema_version__).split('.')
         if int(major) >= 0:
             if int(minor) >= 0:
@@ -289,153 +267,11 @@ class BuildStockBatchBase(object):
         return True
 
     @staticmethod
-    def get_measure_xml(xml_path):
-        tree = ET.parse(xml_path)
-        root = tree.getroot()
-        return root
-
-    @staticmethod
-    def validate_measures_and_arguments(project_file):
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
-        if cfg['stock_type'] != 'residential':  # FIXME: add comstock logic
-            return True
-
-        buildstock_dir = os.path.join(os.path.dirname(project_file), cfg["buildstock_directory"])
-        measures_dir = f'{buildstock_dir}/measures'
-        type_map = {'Integer': int, 'Boolean': bool, 'String': str, 'Double': float}
-
-        measure_names = {
-                        'ResidentialSimulationControls': 'residential_simulation_controls',
-                        'BuildExistingModel': 'baseline',
-                        'SimulationOutputReport': 'simulation_output',
-                        'ServerDirectoryCleanup': None,
-                        'ApplyUpgrade': 'upgrades',
-                        'TimeseriesCSVExport': 'timeseries_csv_export'
-                        }
-        if 'reporting_measures' in cfg.keys():
-            for reporting_measure in cfg['reporting_measures']:
-                measure_names[reporting_measure] = 'reporting_measures'
-
-        error_msgs = ''
-        warning_msgs = ''
-        for measure_name in measure_names.keys():
-            measure_path = os.path.join(measures_dir, measure_name)
-
-            if measure_names[measure_name] in cfg.keys() or \
-                    measure_names[measure_name] == 'residential_simulation_controls':
-                # if they exist in the cfg, make sure they exist in the buildstock checkout
-                if not os.path.exists(measure_path):
-                    error_msgs += f"* {measure_name} does not exist in {buildstock_dir}. \n"
-
-            # check the rest only if that measure exists in cfg
-            if measure_names[measure_name] not in cfg.keys():
-                continue
-
-            # check argument value types for residential simulation controls and timeseries csv export measures
-            if measure_name in ['ResidentialSimulationControls', 'SimulationOutputReport', 'TimeseriesCSVExport']:
-                root = BuildStockBatchBase.get_measure_xml(os.path.join(measure_path, 'measure.xml'))
-                expected_arguments = {}
-                required_args_with_default = {}
-                required_args_no_default = {}
-                for argument in root.findall('./arguments/argument'):
-                    name = argument.find('./name').text
-                    expected_arguments[name] = []
-                    required = argument.find('./required').text
-                    default = argument.find('./default_value')
-                    default = default.text if default is not None else None
-
-                    if required == 'true' and not default:
-                        required_args_no_default[name] = None
-                    elif required == 'true':
-                        required_args_with_default[name] = default
-
-                    if argument.find('./type').text == 'Choice':
-                        for choice in argument.findall('./choices/choice'):
-                            for value in choice.findall('./value'):
-                                expected_arguments[name].append(value.text)
-                    else:
-                        expected_arguments[name] = argument.find('./type').text
-
-                for actual_argument_key in cfg[measure_names[measure_name]].keys():
-                    if actual_argument_key not in expected_arguments.keys():
-                        error_msgs += f"* Found unexpected argument key {actual_argument_key} for "\
-                                      f"{measure_names[measure_name]} in yaml file. The available keys are: " \
-                                      f"{list(expected_arguments.keys())}\n"
-                        continue
-
-                    required_args_no_default.pop(actual_argument_key, None)
-                    required_args_with_default.pop(actual_argument_key, None)
-
-                    actual_argument_value = cfg[measure_names[measure_name]][actual_argument_key]
-                    expected_argument_type = expected_arguments[actual_argument_key]
-
-                    if type(expected_argument_type) is not list:
-                        try:
-                            if type(actual_argument_value) is not list:
-                                actual_argument_value = [actual_argument_value]
-
-                            for val in actual_argument_value:
-                                if not isinstance(val, type_map[expected_argument_type]):
-                                    error_msgs += f"* Wrong argument value type for {actual_argument_key} for measure "\
-                                                  f"{measure_names[measure_name]} in yaml file. Expected type:" \
-                                                  f" {type_map[expected_argument_type]}, got: {val}" \
-                                                  f" of type: {type(val)} \n"
-                        except KeyError:
-                            print(f"Found an unexpected argument value type: {expected_argument_type} for argument "
-                                  f" {actual_argument_key} in measure {measure_name}.\n")
-                    else:  # Choice
-                        if actual_argument_value not in expected_argument_type:
-                            error_msgs += f"* Found unexpected argument value {actual_argument_value} for "\
-                                          f"{measure_names[measure_name]} in yaml file. Valid values are " \
-                                           f"{expected_argument_type}.\n"
-
-                for arg, default in required_args_no_default.items():
-                    error_msgs += f"* Required argument {arg} for measure {measure_name} wasn't supplied. " \
-                                    f"There is no default for this argument.\n"
-
-                for arg, default in required_args_with_default.items():
-                    warning_msgs += f"* Required argument {arg} for measure {measure_name} wasn't supplied. " \
-                                    f"Using default value: {default}. \n"
-
-            elif measure_name in ['ApplyUpgrade']:
-                # For ApplyUpgrade measure, verify that all the cost_multipliers used are correct
-                root = BuildStockBatchBase.get_measure_xml(os.path.join(measure_path, 'measure.xml'))
-                valid_multipliers = set()
-                for argument in root.findall('./arguments/argument'):
-                    name = argument.find('./name')
-                    if name.text.endswith('_multiplier'):
-                        for choice in argument.findall('./choices/choice'):
-                            value = choice.find('./value')
-                            value = value.text if value is not None else ''
-                            valid_multipliers.add(value)
-                invalid_multipliers = Counter()
-                for upgrade_count, upgrade in enumerate(cfg['upgrades']):
-                    for option_count, option in enumerate(upgrade['options']):
-                        for cost_indx, cost_entry in enumerate(option.get('costs', [])):
-                            if cost_entry['multiplier'] not in valid_multipliers:
-                                invalid_multipliers[cost_entry['multiplier']] += 1
-
-                if invalid_multipliers:
-                    error_msgs += "* The following multipliers values are invalid: \n"
-                    for multiplier, count in invalid_multipliers.items():
-                        error_msgs += f"    '{multiplier}' - Used {count} times \n"
-                    error_msgs += f"    The list of valid multipliers are {valid_multipliers}.\n"
-
-        if warning_msgs:
-            logger.warning(warning_msgs)
-
-        if not error_msgs:
-            return True
-        else:
-            logger.error(error_msgs)
-            raise ValidationError(error_msgs)
-
-    @staticmethod
     def validate_options_lookup(project_file):
         """
         Validates that the parameter|options specified in the project yaml file is available in the options_lookup.tsv
         """
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
+        cfg = get_project_configuration(project_file)
         param_option_dict = defaultdict(set)
         buildstock_dir = BuildStockBatchBase.get_buildstock_dir(project_file, cfg)
         options_lookup_path = f'{buildstock_dir}/resources/options_lookup.tsv'
@@ -603,7 +439,7 @@ class BuildStockBatchBase(object):
         Validates that the measures specified in the project yaml file are
         referenced in the options_lookup.tsv
         """
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
+        cfg = get_project_configuration(project_file)
         measure_dirs = set()
         buildstock_dir = BuildStockBatchBase.get_buildstock_dir(project_file, cfg)
         options_lookup_path = f'{buildstock_dir}/resources/options_lookup.tsv'
@@ -658,7 +494,7 @@ class BuildStockBatchBase(object):
         """
         Checks if the reference_scenario mentioned in an upgrade points to a valid upgrade
         """
-        cfg = BuildStockBatchBase.get_project_configuration(project_file)
+        cfg = get_project_configuration(project_file)
 
         # collect all upgrade_names
         upgrade_names = set()
