+++ buildstockbatch/test/test_postprocessing.py
@@ -10,6 +10,7 @@ import shutil
 
 from buildstockbatch import postprocessing
 from buildstockbatch.base import BuildStockBatchBase
+from buildstockbatch.utils import get_project_configuration
 from unittest.mock import patch
 
 
@@ -48,7 +49,7 @@ def test_report_additional_results_csv_columns(basic_residential_project_file):
     with gzip.open(sim_out_dir / 'results_job0.json.gz', 'wt', encoding='utf-8') as f:
         json.dump(dpouts2, f)
 
-    cfg = BuildStockBatchBase.get_project_configuration(project_filename)
+    cfg = get_project_configuration(project_filename)
 
     postprocessing.combine_results(fs, results_dir, cfg, do_timeseries=False)
 
@@ -67,7 +68,7 @@ def test_empty_results_assertion(basic_residential_project_file, capsys):
     results_dir = pathlib.Path(results_dir)
     sim_out_dir = results_dir / 'simulation_output'
     shutil.rmtree(sim_out_dir)  # no results
-    cfg = BuildStockBatchBase.get_project_configuration(project_filename)
+    cfg = get_project_configuration(project_filename)
 
     with pytest.raises(ValueError, match=r'No simulation results found to post-process'):
         assert postprocessing.combine_results(fs, results_dir, cfg, do_timeseries=False)
